Renamed the css function to resource to be more generally available should the user have a mime type that is not currently supported.
Added functions named: font, css, js, and img. All of which check the
file extension of the file that we are working with and return the
correct MIME for that file.

Currently jpg/jpeg, gif, png, bmp, ico, and tif/tiff images are
supported, as well as most possible web fonts.